requirements_final.txt
absl-py==2.3.0
art==6.5
asttokens==2.4.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
efficientnet_pytorch==0.7.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
grpcio==1.73.0
huggingface-hub==0.23.5
idna==3.10
imageio==2.37.0
importlib_metadata==8.4.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==6.31.1
psutil==6.0.0
pure_eval==0.2.3
pycm==4.3
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.4
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.4.2
scikit-multilearn==0.2.0
scipy==1.13.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.9.16
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.20.0
